requirements.txt
certifi==2022.12.7
click==8.1.3
clickhouse-connect==0.5.16
clickhouse-driver==0.2.5
Flask==2.2.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
lz4==4.3.2
MarkupSafe==2.1.2
numpy==1.24.2
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
spectree==1.1.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
vertica-python==1.3.1
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
zstandard==0.20.0
